+++ finetune/finetune_power.py
@@ -17,7 +17,7 @@ from torch import nn
 import os
 from random import randrange
 from torch.utils import data
-from models.pangu_power_sample import test, train
+from wind_fusion.pangu_pytorch.models.train_power import test, train
 from models.pangu_power import (
     PanguPowerPatchRecovery,
     PanguPowerConvSigmoid,
